requirements.txt
PySide6
Pillow
mutagen
pytest
